+++ Code/DataNode.go
@@ -175,7 +175,10 @@ func remover(){
     panic(err)
   }
   for i:=1;i<len(files);i++{
-    os.Remove(files[i])
+    wea:=strings.Split(files,"/")
+    if(wea[len(wea)-1]!=""){
+      os.Remove(files[i])
+    }
   }
 }
 func main(){
